Metadata: Enable Multi-author documents and Increase entity richness for Authors
Make it possible to list multiple authors on a given document.
Also, increase the richness of authors as entities to increase interactiveness in a document (for instance, creating explorable, clickable names on front-end visualizations of data, working towards the possibility of uniquely identifying author edits, etc.).
This could be expanded out in different ways, but this is a proposal to create a basic initial effort here. One potential expansion would be to add a field for a given organization's site.

Also, adding a creationDate to metadata.